requirements.txt
absl-py
certifi==2021.10.8
charset-normalizer==2.0.8
cloudpickle
grpcio
gym==0.10.5
idna==3.3
importlib-metadata
Markdown
numpy
olefile
Pillow==5.4.1
protobuf==3.17.2
pyglet==1.5.21
PyYAML==6.0
requests==2.26.0
six
tensorboard==1.15.0
#torch==1.8.0
#torchaudio==0.8.0
#torchvision==0.9.0
typing_extensions
urllib3==1.26.7
webencodings==0.5.1
Werkzeug
wincertstore==0.2
zipp
